Migrate featuredVendors logic from Routes to Controller
